requirements.txt
mysql-connector==2.2.9
mysql-connector-python==9.1.0
python-dotenv==1.0.0
SQLAlchemy==2.0.35
mysqlclient==2.2.6